requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.11.2
pydantic-settings==2.1.0

# LLM integrations
openai>=1.52.0
anthropic>=0.40.0
groq>=0.12.0
boto3>=1.34.0
botocore>=1.34.0
transformers==4.35.2
langchain==0.0.339
llama-index==0.9.13

# Temporal workflow
temporalio==1.14.1

# Vector databases
qdrant-client==1.14.3
weaviate-client==3.25.3

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1

# Message queue
aiokafka==0.10.0
confluent-kafka==2.3.0

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-jaeger-thrift==1.21.0
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7
sentry-sdk==1.39.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
slowapi==0.1.9
PyJWT>=2.9.0

# Clerk Authentication
clerk-backend-api==3.0.3

# Utilities
httpx>=0.28.1
tenacity==8.2.3
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7

# Code execution sandbox
docker==7.1.0
kubernetes==28.1.0

# Validation
pylint==3.0.2
black==23.11.0
mypy==1.7.1
ruff==0.1.6
bandit==1.7.5
safety==3.2.8

# Machine Learning (for validation confidence engine)
scikit-learn==1.3.2
numpy==1.24.3

# Production-Grade Quality Assurance
pytest==7.4.3
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
bandit==1.7.5
coverage==7.3.2
radon==6.0.1
xenon==0.9.1

# Performance Testing & Monitoring
locust==2.17.0
memory-profiler==0.61.0
py-spy==0.3.14

# Security Testing
safety==3.2.8
semgrep==1.45.0
pip-audit==2.6.1

# Code Quality & Static Analysis
flake8==6.1.0
isort==5.12.0
autoflake==2.2.1
pre-commit==3.5.0

# Social Media APIs (for marketing auto-publish)
tweepy==4.14.0
linkedin-api==2.1.1
